refactor: rename `actors` to `actions`
