+++ internal/actions/actions.go
@@ -1,22 +1,22 @@
-package actors
+package actions
 
 import (
 	"io"
 
-	"github.com/nobe4/gh-not/internal/actors/debug"
-	"github.com/nobe4/gh-not/internal/actors/done"
-	"github.com/nobe4/gh-not/internal/actors/hide"
-	"github.com/nobe4/gh-not/internal/actors/open"
-	"github.com/nobe4/gh-not/internal/actors/pass"
-	"github.com/nobe4/gh-not/internal/actors/print"
-	"github.com/nobe4/gh-not/internal/actors/read"
+	"github.com/nobe4/gh-not/internal/actions/debug"
+	"github.com/nobe4/gh-not/internal/actions/done"
+	"github.com/nobe4/gh-not/internal/actions/hide"
+	"github.com/nobe4/gh-not/internal/actions/open"
+	"github.com/nobe4/gh-not/internal/actions/pass"
+	"github.com/nobe4/gh-not/internal/actions/print"
+	"github.com/nobe4/gh-not/internal/actions/read"
 	"github.com/nobe4/gh-not/internal/gh"
 	"github.com/nobe4/gh-not/internal/notifications"
 )
 
-type ActorsMap map[string]Actor
+type ActionsMap map[string]Actor
 
-func Map(client *gh.Client) ActorsMap {
+func Map(client *gh.Client) ActionsMap {
 	return map[string]Actor{
 		"pass":  &pass.Actor{},
 		"debug": &debug.Actor{},
