+++ internal/actions/hide/hide.go
